refactor: Merge ReservationUnitSettings component files
